+++ Changelog.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file since v2.0.0
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.2.0] - 2020-10-05
+
+### Changed
+
+- Updated dependencies and locked minor version for better stability.
+- Parameter name and usage from `$elasticsearchVersion` to `$elasticsearchHost` in the Magneto-related commands.
+
+### Removed
+
+- Removed option `--mysql-container` from the command `magento:test-module-install` to dynamically find current linked MySQL container.
+
 
 ## [2.1.0] - 2020-08-04